pyproject.toml
[project]
name = "inferflow"
version = "1.0.0"
description = ""
authors = [
    {name = "BoChenSHEN",email = "6goddddddd@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<3.15"
dependencies = [
    "torch (>=2.1.0,<3.0.0)",
    "torchvision (>=0.24.1,<0.25.0)",
    "numpy (>=1.23,<1.27)",
    "opencv-python (>=4.5.5,<5.0)",
    "pillow (>=10.0.0,<13.0.0)",
]


[project.optional-dependencies]
test = ["pytest (>=9.0.1,<10.0.0)", "pytest-cov (>=7.0.0,<8.0.0)", "mypy (>=1.19.0,<2.0.0)", "pylint (>=4.0.4,<5.0.0)"]
dev = ["ruff (>=0.14.7,<0.15.0)", "pre-commit (>=4.5.0,<5.0.0)"]
docs = [
        "mkdoc (>=0.1,<0.2)",
    "mkdocs-material (>=9.7.0,<10.0.0)",
    "mkdocs-minify-plugin (>=0.8.0,<0.9.0)",
    "mkdocstrings[python] (>=1.0.0,<2.0.0)",
    "pymdown-extensions (>=10.17.2,<11.0.0)"
]
[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
